Made .punish do stuff

+++ bot/moderation.py
@@ -39,16 +39,95 @@ class moderation(commands.Cog):
     @commands.command(aliases=["pu"])
     @commands.check(Checks.mod)
     async def punish(self, ctx, users: commands.Greedy[discord.Member], punishment: typing.Optional[parse_punishment] = 0, duration: commands.Greedy[parse_time] = [], *, reason):
+        if punishment == 3 and duration != 0:
+            raise commands.BadArgument
+        
         await ctx.message.delete()
+
+        ref = (f"https://discord.com/channels/{ctx.guild.id}/{ctx.channel.id}/" +
+               str(ctx.message.reference.message_id if ctx.message.reference is not None else ctx.channel.last_message_id))
 
         id = await self.bot.db.fetchval("SELECT incident_index FROM guilds WHERE id = $1", ctx.guild.id)
         await self.bot.db.execute("UPDATE guilds SET incident_index = incident_index + 1 WHERE id = $1", ctx.guild.id)
         
         await self.bot.db.execute("INSERT INTO incidents (guild, id, moderator, users, type_, time_, expires, comment, ref)" +
                                   "VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)", ctx.guild.id, id, ctx.author.id, [user.id for user in users],
-                                  punishment, time.time(), time.time() + sum(duration), reason,
-                                  f"https://discord.com/channels/{ctx.guild.id}/{ctx.channel.id}/" +
-                                  str(ctx.message.reference.message_id if ctx.message.reference is not None else ctx.channel.last_message_id))
+                                  punishment, time.time(), time.time() + sum(duration), reason, ref)
+
+        delta = []
+        temp = sum(duration)
+        if temp // (60**2 * 24 * 365) > 0:
+            delta.append(f"{temp // (60**2 * 24 * 365)} years")
+            temp %= 60**2 * 24 * 365
+        if temp // (60**2 * 24 * 30) > 0:
+            delta.append(f"{temp // (60**2 * 24 * 30)} months")
+            temp %= 60**2 * 24 * 30
+        if temp // (60**2 * 24) > 0:
+            delta.append(f"{temp // (60**2 * 24)} days")
+            temp %= 60**2 * 24
+        if temp // 60**2 > 0:
+            delta.append(f"{temp // 60**2} hours")
+            temp %= 60**2
+        if temp // 60 > 0:
+            delta.append(f"{temp // 60} minutes")
+            temp %= 60
+        if temp > 0:
+            delta.append(f"{temp} seconds")
+        del temp
+        if len(delta) == 0:
+            delta = ["Forever"]
+        delta = ", ".join(delta)
+        
+        async def log(type):
+            colours = {
+                "Warn": 0xfc9003,
+                "Mute": 0xc74c00,
+                "Kick": 0xff0077,
+                "Ban": 0xff0000
+            }
+
+            embed = discord.Embed(title=f"Incident #{id}", colour=colours[type])
+            embed.set_author(name=ctx.author.display_name, icon_url=ctx.author.avatar_url)
+            embed.add_field(name="Members involved", value=", ".join([member.mention for member in users]))
+            embed.add_field(name="Punishemnt", value=type)
+            if punishment in (1, 2, 4):
+                embed.add_field(name="Duration", value=delta)
+            embed.add_field(name="Reason", value=f"{reason} ([ref]({ref}))", inline=False)
+
+            for channel in [ctx.guild.get_channel(record["id"]) for record in await self.bot.db.fetch("SELECT id FROM channels WHERE log_type = 'moderation' AND guild_id = $1", ctx.guild.id)]:
+                await channel.send(embed=embed)
+
+        async def none():
+            pass
+
+        async def warn():
+            await log("Warn")
+
+        async def mute():
+            role = ctx.guild.get_role(await self.bot.db.fetchval("SELECT mute_role FROM guilds WHERE id = $1", ctx.guild.id))
+            for user in users:
+                await user.add_roles(role, reason=f"Duration: {delta}. Incident #{id}: {reason}")
+            await log("Mute")
+
+        async def kick():
+            for user in users:
+                await user.kick(reason=f"Incident #{id}: {reason}")
+            await log("Kick")
+
+        async def ban():
+            for user in users:
+                await user.ban(reason=f"Duration: {delta}. Incident #{id}: {reason}", delete_message_days=0)
+            await log("Ban")
+
+        punishments = {
+            0: none,
+            1: warn,
+            2: mute,
+            3: kick,
+            4: ban
+        }
+
+        await punishments[punishment]()
         
 
 def setup(bot):
